✨ feat: 파일 검증부 로직 개선
